+++ tests/unit_tests/backtest/test_backtest_exchange_l2_mbp.py
@@ -37,7 +37,6 @@ from nautilus_trader.model.enums import BookType
 from nautilus_trader.model.enums import OMSType
 from nautilus_trader.model.enums import OrderSide
 from nautilus_trader.model.enums import OrderStatus
-from nautilus_trader.model.enums import VenueType
 from nautilus_trader.model.identifiers import Venue
 from nautilus_trader.model.objects import Money
 from nautilus_trader.model.objects import Price
@@ -106,7 +105,6 @@ class TestL2OrderBookExchange:
 
         self.exchange = SimulatedExchange(
             venue=SIM,
-            venue_type=VenueType.ECN,
             oms_type=OMSType.HEDGING,
             account_type=AccountType.MARGIN,
             base_currency=USD,