+++ tests/unit_tests/live/test_live_risk_engine.py
@@ -30,7 +30,6 @@ from nautilus_trader.model.commands.trading import SubmitOrder
 from nautilus_trader.model.currencies import USD
 from nautilus_trader.model.enums import AccountType
 from nautilus_trader.model.enums import OrderSide
-from nautilus_trader.model.enums import VenueType
 from nautilus_trader.model.identifiers import ClientId
 from nautilus_trader.model.identifiers import StrategyId
 from nautilus_trader.model.identifiers import TraderId
@@ -115,7 +114,6 @@ class TestLiveRiskEngine:
 
         self.exec_client = MockExecutionClient(
             client_id=ClientId("SIM"),
-            venue_type=VenueType.ECN,
             account_id=TestStubs.account_id(),
             account_type=AccountType.MARGIN,
             base_currency=USD,